+++ src/main/java/pl/plajer/murdermystery/Main.java
@@ -197,6 +197,7 @@ public class Main extends JavaPlugin {
           for (PotionEffect pe : player.getActivePotionEffects()) {
             player.removePotionEffect(pe.getType());
           }
+          player.setWalkSpeed(0.2f);
         }
       }
       arena.teleportAllToEndLocation();