+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 * Fixed loading arena values (Some values had a wrong file location)
 * Added party dependency, you can join with more players to one arena
 * Fixed murder gets sword as spectator
+* Changed murder speed boost from potion to walkspeed to hide particles
 
 ### 1.4.1 Release (05.12.2019 - 13.12.2019) (by Tigerpanzer)
 * Added a new MOTD Manager in the bungee.yml (Now you can define the states on yourself)